Applying Discount Feature Added in Game and Discount Service and POM version Updated

+++ RetroRealm-GameService/src/main/java/com/Batman/dto/game/BaseGameResponse.java
@@ -15,7 +15,7 @@ public class BaseGameResponse {
 	private Double gamePrice;
 	private Double gameVersion;
 	private LocalDate gameReleasedDate;
-	private Integer gameDiscount;
+	private Double gameDiscount;
 	
 	private Set<GameGenre> gameGenere;
 }
